fix(world): apply remaining code-review cleanups
- ring.rs: index via (tick % RING_SIZE) in u64 before the usize cast, so
  ticks above 2^32 don't alias to the same slot on 32-bit targets.
- Entities::keys returns impl Iterator<Item = EntityId> instead of leaking
  the internal hecs::Entity in the public signature.
- Drop the write-only client-side Entities.properties store (and its
  maintenance in apply/remove/retain): it was never read. The client now
  ignores property deltas (HP) until a HUD needs them; the server still
  sends them.
- CLAUDE.md: correct the claim of SnapshotRing unit tests (none exist;
  the ring now lives in ring.rs, keyed by Tick).

+++ crates/blackflower-authority/src/ring.rs
@@ -18,14 +18,15 @@ impl Default for SnapshotRing {
 
 impl SnapshotRing {
     pub fn insert(&mut self, tick: Tick, snapshot: WorldSnapshot) {
-        self.entries[tick.as_u64() as usize % RING_SIZE] = Some((tick, snapshot));
+        self.entries[(tick.as_u64() % RING_SIZE as u64) as usize] = Some((tick, snapshot));
     }
 
     pub fn get(&self, tick: Tick) -> Option<&WorldSnapshot> {
         if tick == Tick::ZERO {
             return None;
         }
-        let (stored_tick, snapshot) = self.entries[tick.as_u64() as usize % RING_SIZE].as_ref()?;
+        let (stored_tick, snapshot) =
+            self.entries[(tick.as_u64() % RING_SIZE as u64) as usize].as_ref()?;
         (*stored_tick == tick).then_some(snapshot)
     }
 }
